gpio.c bug fixed

+++ linux/gpio.c
@@ -61,7 +61,7 @@ int mb_stop()
 
 
 
-int set_gpio_to_hi(uint8_t gpio_pin)
+int set_gpio_to_hi(uint16_t gpio_pin)
 {
   int status, gpio_val;
   char gpio_str[4];
@@ -88,7 +88,7 @@ int set_gpio_to_hi(uint8_t gpio_pin)
   return 0;
 }
 
-int set_gpio_to_low(uint8_t gpio_pin)
+int set_gpio_to_low(uint16_t gpio_pin)
 {
   int status, gpio_val;
   char gpio_str[4];
@@ -115,7 +115,7 @@ int set_gpio_to_low(uint8_t gpio_pin)
   return 0;
 }
 
-int init_gpio (uint8_t gpio_pin)
+int init_gpio (uint16_t gpio_pin)
 {
   int gpio_exp, gpio_direction;
   char gpio_str[4];
@@ -147,7 +147,7 @@ int init_gpio (uint8_t gpio_pin)
   return 0;
 }
 
-int close_gpio (uint8_t gpio_pin)
+int close_gpio (uint16_t gpio_pin)
 {
   int gpio_unexp;
   char gpio_str[4];
